requirements.txt
certifi==2020.6.20
cffi==1.14.1
cryptography==3.0
decorator==4.4.2
future==0.18.2
peewee==3.13.3
pycparser==2.20
python-dotenv==0.14.0
python-telegram-bot==12.8
six==1.15.0
tornado==6.0.4
